Fixed resizeTable function
1) In resizeTable function when copying the table, the key was
not initialized. Probably because I used the keys array field in
the ht struct before, but then erased it and replaced it with the
ht_get_keys function without updating this function.
The fix was simple, just get the keys with the ht_get_keys function
and use it to loop over the table.

2) Also in the resizeTable function, I only free the ht_copy so a
memmory leak occured.
The fix was to free the ht_copy->arr, and the keys array I used to
loop over the table.
After the fix valgrind confirmed no leaks were found.