ht\_insert returns int to determine update/new
1) Changed the return value of ht\_insert function to int.
The function can return one of two newly defined macros:
HT\_INSERT\_EMPTY and HT\_INSERT\_UPDATE which can tell the user
if the key already existed so the entry has been updated or the key
is new and a new entry has been added.

2) The static functions declarations were transfered from the header
file to the source file, so it will not pollute the namespace when used
as a library.